認可の追加、記事の執筆者のみ編集削除ボタンが見れるように。あとbookmarks tablesのマイグレーションファイル作成。

+++ src/app/Http/Controllers/ArticleController.php
@@ -73,6 +73,8 @@ class ArticleController extends Controller
      */
     public function edit(Article $article)
     {
+        $this->authorize($article);
+
         $data = ['article' => $article];
         return view('articles.edit', $data);
     }
@@ -86,6 +88,8 @@ class ArticleController extends Controller
      */
     public function update(Request $request, Article $article)
     {
+        $this->authorize($article);
+
         $this->validate($request, [
             'title' => 'required|max:255',
             'body' => 'required',
@@ -104,6 +108,8 @@ class ArticleController extends Controller
      */
     public function destroy(Article $article)
     {
+        $this->authorize($article);
+
         $article->delete();
         return redirect(route('articles.index'));
     }
